fix error cast on ReadFrom return in sendAndRead
The sendAndRead function is casting the result of ReadFrom on its
underlying connection to a *net.OpError, which seemingly would work if
the error that came back was from the net package. However, it appears
that the error now comes from the github.com/mdlayher/raw package. The
error does conform to the net.Error interface, but is not a *net.OpError
so the cast will fail every time. This commit changes the cast to a
net.Error interface.

With the previous behavior u-root/cmds/dhclient would fail with an "i/o
timeout" every time. With this change dhclient works.